refactor: lambdas and add email to subscribe sns

+++ lambdas/generate_file/lambda_handler.py
@@ -7,15 +7,17 @@ import boto3
 import logging
 from faker import Faker
 
+# Configuração do logger
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger()
 
 def generate_random_file(max_lines=500):
+    """Gera um arquivo com dados aleatórios usando Faker."""
     fake = Faker('pt_BR')
     num_lines = random.randint(50, max_lines)
     filename = f"{uuid.uuid4()}.txt"
-    filepath = f"/dev/shm/{filename}"
-    
+    filepath = f"/tmp/{filename}"  # Diretório temporário padrão no AWS Lambda
+
     try:
         with open(filepath, 'w', encoding='utf-8') as f:
             for _ in range(num_lines):
@@ -36,16 +38,31 @@ def generate_random_file(max_lines=500):
         raise
 
 def lambda_handler(event, context):
+    """Handler da função Lambda."""
     s3_client = boto3.client('s3')
     BUCKET_NAME = os.environ.get('BUCKET_NAME')
-    
+
+    if not BUCKET_NAME:
+        logger.error("A variável de ambiente BUCKET_NAME não foi definida.")
+        return {
+            'statusCode': 500,
+            'body': json.dumps({'message': 'Erro de configuração: BUCKET_NAME não está definido.'})
+        }
+
     try:
         start_time = time.time()
         filename, filepath, num_lines = generate_random_file()
-        s3_client.upload_file(filepath, BUCKET_NAME, filename)
-        os.remove(filepath)
-        
-        logger.info(f"Arquivo {filename} gerado e enviado. Linhas: {num_lines}")
+
+        try:
+            s3_client.upload_file(filepath, BUCKET_NAME, filename)
+            logger.info(f"Arquivo {filename} com {num_lines} linhas gerado e enviado em {time.time() - start_time:.2f} segundos.")
+        except boto3.exceptions.S3UploadFailedError as e:
+            logger.error(f"Erro no upload para o S3: {e}")
+            raise
+        finally:
+            if os.path.exists(filepath):
+                os.remove(filepath)
+
         return {
             'statusCode': 200,
             'body': json.dumps({
